docs: add route refactor progress file
